EDA_Agent/requirements.txt
matplotlib
numpy
openai
pandas
python-dotenv
seaborn
streamlit
langchain
langchain-experimental
langchain-openai
langchain-anthropic
tabulate